클래스 분리 작업 진행
