+++ utils/elements.py
@@ -1,31 +1,6 @@
-from utils import converter
-from constants import REQUEST
 from constants import REQUEST as RQ
 
 import hashlib
-
-
-class FileMeta:
-    UID_LENGTH = 64
-    DIR_LENGTH = 200
-    KEY_LENGTH = 100
-
-    def __init__(self, tokenData, keyDirData):
-        decoded = converter.jwtTokenDecoder(tokenData[RQ.TOKEN])
-        email = decoded['email']
-        provider = decoded['provider']
-
-        self.uid = getUid(email, provider)
-        self.dir = keyDirData[RQ.DIR]
-        self.key = keyDirData[RQ.KEY]
-
-    @property
-    def data(self):
-        return {
-            RQ.UID: self.uid,
-            RQ.DIR: self.dir,
-            RQ.KEY: self.key
-        }
 
 
 def getUid(email, provider):
@@ -37,7 +12,7 @@ def getListObject(s3Client, s3Bucket, uid):  # 해당 userID를 가진 컨텐츠
     try:
         return s3Client.list_objects_v2(Bucket=s3Bucket, Prefix=uid)['Contents']
     except KeyError:
-        return REQUEST.DEFAULT_OBJECT_LIST
+        return RQ.DEFAULT_OBJECT_LIST
 
 
 def getContents(listObject):
